modify search and schedule view
