+++ src/view/search_product/SearchProductView.java
@@ -4,10 +4,7 @@ import entity.product.Product;
 import entity.user.User;
 import interface_adapter.login.ViewLoginPageController;
 import interface_adapter.logout.LogOutController;
-import interface_adapter.search_product.GetSearchPageController;
-import interface_adapter.search_product.SearchProductByNameController;
-import interface_adapter.search_product.SearchProductByTagController;
-import interface_adapter.search_product.SearchProductViewModel;
+import interface_adapter.search_product.*;
 import interface_adapter.shopping_cart.ShoppingCartController;
 import interface_adapter.signup.ViewSignupPageController;
 import interface_adapter.view_product.ViewProductController;
@@ -40,6 +37,7 @@ public class SearchProductView extends JPanel implements ActionListener, Propert
     private JTextField searchBox;
     private JButton searchButton;
     private final String[] tags = {"Tag1", "Tag2", "Tag3"}; // to be decided later
+    AllProductsPanel allProductsPanel;
 
     public SearchProductView(SearchProductByNameController searchByNameController,
                              SearchProductByTagController searchByTagController,
@@ -110,67 +108,70 @@ public class SearchProductView extends JPanel implements ActionListener, Propert
 
         // products display starts here
         ArrayList<Product> products = viewModel.getState().getProducts();
+        allProductsPanel = new AllProductsPanel(products, viewModel, viewProductController);
+        this.add(allProductsPanel);
 
-        int _i = 0;
-        List<JPanel> listProductPanels = new ArrayList<>();
-
-        for (Product product: products) {
-            if (product.getState() == 0) {
-
-                Image image = product.getImage();
-                JLabel paneledImage = new JLabel(new ImageIcon(image));
-                JLabel productTitle = new JLabel(product.getTitle());
-
-                JLabel productPrice = new JLabel(String.valueOf(product.getPrice()));
-
-                JButton viewButton = new JButton(product.getTitle());
-                // dimension set as this for now but will likely get changed later
-                viewButton.setPreferredSize(new Dimension(100, 50));
-                viewButton.addActionListener(
-                        new ActionListener() {
-                            public void actionPerformed(ActionEvent event) {
-                                if (event.getSource().equals(viewButton)) {
-                                    User user = viewModel.getState().getUser();
-                                    try {
-                                        viewProductController.execute(product, user);
-                                    } catch (SQLException e) {
-                                        throw new RuntimeException(e); //Revisit this in case of bug in viewing a product
-                                    }
-
-                                }
-                            }
-                        }
-                );
-
-                view.search_product.ProductPanel productPanel = new view.search_product.ProductPanel(
-                        paneledImage, productTitle, productPrice, viewButton
-                );
-
-                // Above created one panel for image
-
-                if (_i % 3 == 0) {
-                    listProductPanels = new ArrayList<>();
-
-                }
-                listProductPanels.add(productPanel);
-
-                if (_i % 3 == 2) {
-                    view.search_product.HorizontalLayoutPanel horizontalLayoutPanel = new view.search_product.HorizontalLayoutPanel(
-                            listProductPanels
-                    );
-                    this.add(horizontalLayoutPanel);
-                } else if (_i + 1 == products.size()) {
-                    view.search_product.HorizontalLayoutPanel horizontalLayoutPanel = new view.search_product.HorizontalLayoutPanel(
-                            listProductPanels
-                    );
-                    this.add(horizontalLayoutPanel);
-                }
-
-                _i++;
-
-            }
-        }
-        // Products panel ends here
+        // Wrong implementation of SearchProductView, left for reference
+//        int _i = 0;
+//        List<JPanel> listProductPanels = new ArrayList<>();
+//
+//        for (Product product: products) {
+//            if (product.getState() == 0) {
+//
+//                Image image = product.getImage();
+//                JLabel paneledImage = new JLabel(new ImageIcon(image));
+//                JLabel productTitle = new JLabel(product.getTitle());
+//
+//                JLabel productPrice = new JLabel(String.valueOf(product.getPrice()));
+//
+//                JButton viewButton = new JButton(product.getTitle());
+//                // dimension set as this for now but will likely get changed later
+//                viewButton.setPreferredSize(new Dimension(100, 50));
+//                viewButton.addActionListener(
+//                        new ActionListener() {
+//                            public void actionPerformed(ActionEvent event) {
+//                                if (event.getSource().equals(viewButton)) {
+//                                    User user = viewModel.getState().getUser();
+//                                    try {
+//                                        viewProductController.execute(product, user);
+//                                    } catch (SQLException e) {
+//                                        throw new RuntimeException(e); //Revisit this in case of bug in viewing a product
+//                                    }
+//
+//                                }
+//                            }
+//                        }
+//                );
+//
+//                view.search_product.ProductPanel productPanel = new view.search_product.ProductPanel(
+//                        paneledImage, productTitle, productPrice, viewButton
+//                );
+//
+//                // Above created one panel for image
+//
+//                if (_i % 3 == 0) {
+//                    listProductPanels = new ArrayList<>();
+//
+//                }
+//                listProductPanels.add(productPanel);
+//
+//                if (_i % 3 == 2) {
+//                    view.search_product.HorizontalLayoutPanel horizontalLayoutPanel = new view.search_product.HorizontalLayoutPanel(
+//                            listProductPanels
+//                    );
+//                    this.add(horizontalLayoutPanel);
+//                } else if (_i + 1 == products.size()) {
+//                    view.search_product.HorizontalLayoutPanel horizontalLayoutPanel = new view.search_product.HorizontalLayoutPanel(
+//                            listProductPanels
+//                    );
+//                    this.add(horizontalLayoutPanel);
+//                }
+//
+//                _i++;
+//
+//            }
+//        }
+//        // Products panel ends here
 
 
     }
@@ -184,6 +185,10 @@ public class SearchProductView extends JPanel implements ActionListener, Propert
 
     @Override
     public void propertyChange(PropertyChangeEvent evt) {
+        SearchProductState searchProductState = (SearchProductState) evt.getNewValue();
+
+        ArrayList<Product> products = searchProductState.getProducts();
+        allProductsPanel = new AllProductsPanel(products, viewModel, viewProductController);
 
     }
